Add color to appropriate places, break out stylesheets to match components
