+++ styles/modules/PolicyEvaluationDetails.module.scss
@@ -17,38 +17,29 @@
 @import "styles/constants";
 @import "styles/mixins";
 
-.container {
-  width: 100%;
-  height: 100%;
-}
 
-.resourceEvaluationCard {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  margin: 1rem auto;
+.policyEvaluationCard {
   padding: 1rem;
-
-  @include tabletAndLarger {
-    padding: 1rem 2rem;
-  }
+  margin: 1rem auto;
 }
 
-.resourceEvaluationCardHeader, .policyEvaluationCardHeader {
+.policyEvaluationCardHeader {
   display: flex;
   flex-direction: row;
   justify-content: space-between;
   align-items: center;
 }
 
-.policyEvaluationsContainer {
-  width: 95%;
-  margin: 1rem auto;
-}
+.policyDetails {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  align-items: center;
+  width: fit-content;
 
-.policyEvaluationCard {
-  padding: 1rem;
-  margin: 1rem auto;
+  > p {
+    margin: 0 1rem;
+  }
 }
 
 .violations {
@@ -64,26 +55,37 @@
   }
 }
 
+.violationResult {
+  text-transform: uppercase;
+  font-weight: 600;
+}
+
 .darkTheme {
-  .resourceEvaluationCard {
-    background-color: $DT_BACKGROUND_MEDIUM;
+ &.policyEvaluationCard {
+    background-color: $DT_BACKGROUND_DARK;
     box-shadow: $DT_BOX_SHADOW;
   }
 
-  .policyEvaluationCard {
-    background-color: $DT_BACKGROUND_DARK;
-    box-shadow: $DT_BOX_SHADOW;
+  .pass {
+    color: $BRIGHT_GREEN;
+  }
+
+  .fail {
+    color: $HOT_RED;
   }
 }
 
 .lightTheme {
-  .resourceEvaluationCard {
-    background-color: $LT_BACKGROUND_LIGHT;
+  &.policyEvaluationCard {
+    background-color: $LT_BACKGROUND_DARK;
     box-shadow: $LT_BOX_SHADOW;
   }
 
-  .policyEvaluationCard {
-    background-color: $LT_BACKGROUND_DARK;
-    box-shadow: $LT_BOX_SHADOW;
+  .pass {
+    color: $GREEN;
+  }
+
+  .fail {
+    color: $HOT_RED;
   }
 }
