+++ styles/modules/ResourceEvaluation.module.scss
@@ -1,0 +1,71 @@
+/**
+ * Copyright 2021 The Rode Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@import "styles/constants";
+@import "styles/mixins";
+
+.resourceEvaluationCard {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  margin: 1rem auto;
+  padding: 1rem;
+
+  @include tabletAndLarger {
+    padding: 1rem 2rem;
+  }
+}
+
+.resourceEvaluationCardHeader {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.policyEvaluationsContainer {
+  margin-top: 1rem;
+}
+
+.darkTheme {
+  &.resourceEvaluationCard {
+    background-color: $DT_BACKGROUND_MEDIUM;
+    box-shadow: $DT_BOX_SHADOW;
+  }
+
+  .pass {
+    color: $BRIGHT_GREEN;
+  }
+
+  .fail {
+    color: $HOT_RED;
+  }
+}
+
+.lightTheme {
+  &.resourceEvaluationCard {
+    background-color: $LT_BACKGROUND_LIGHT;
+    box-shadow: $LT_BOX_SHADOW;
+  }
+
+  .pass {
+    color: $GREEN;
+  }
+
+  .fail {
+    color: $HOT_RED;
+  }
+}
